requirements.txt
django>=4.1.7,<4.1.8
djangorestframework>=3.14,<3.15
drf_spectacular>=0.26.1,<0.27
